+++ spec/core/HBool.spec.ts
@@ -4,7 +4,7 @@
 
 import { HBool } from '../../src/core/HBool'
 import { HGrid } from '../../src/core/grid/HGrid'
-import { HList } from '../../src/core/HList'
+import { HList } from '../../src/core/list/HList'
 import { HDict } from '../../src/core/dict/HDict'
 import { Kind } from '../../src/core/Kind'
 import { TEXT_ENCODER } from '../../src/core/HVal'
