+++ spec/core/HDateTime.spec.ts
@@ -5,7 +5,7 @@
 import { HDateTime } from '../../src/core/HDateTime'
 import { Kind } from '../../src/core/Kind'
 import { HGrid } from '../../src/core/grid/HGrid'
-import { HList } from '../../src/core/HList'
+import { HList } from '../../src/core/list/HList'
 import { HDict } from '../../src/core/dict/HDict'
 import { TEXT_ENCODER } from '../../src/core/HVal'
