+++ spec/core/HNamespace.spec.ts
@@ -9,7 +9,7 @@ import { HSymbol } from '../../src/core/HSymbol'
 import { ZincReader } from '../../src/core/ZincReader'
 import { HRef } from '../../src/core/HRef'
 import { HStr } from '../../src/core/HStr'
-import { HList } from '../../src/core/HList'
+import { HList } from '../../src/core/list/HList'
 import { HMarker } from '../../src/core/HMarker'
 import { HNum } from '../../src/core/HNum'
 import { Kind } from '../../src/core/Kind'
