+++ spec/core/HNum.spec.ts
@@ -11,7 +11,7 @@ import {
 } from '../../src/core/HNum'
 import { Kind } from '../../src/core/Kind'
 import { HGrid } from '../../src/core/grid/HGrid'
-import { HList } from '../../src/core/HList'
+import { HList } from '../../src/core/list/HList'
 import { HDict } from '../../src/core/dict/HDict'
 import { HUnit } from '../../src/core/HUnit'
 import {
