+++ spec/core/TrioReader.spec.ts
@@ -13,7 +13,7 @@ import { HDate } from '../../src/core/HDate'
 import { HDict } from '../../src/core/dict/HDict'
 import { HSymbol } from '../../src/core/HSymbol'
 import { HGrid } from '../../src/core/grid/HGrid'
-import { HList } from '../../src/core/HList'
+import { HList } from '../../src/core/list/HList'
 import { Kind } from '../../src/core/Kind'
 import '../matchers'
 import '../customMatchers'
