+++ spec/core/Util.spec.ts
@@ -28,7 +28,7 @@ import { HRemove } from '../../src/core/HRemove'
 import { HCoord } from '../../src/core/HCoord'
 import { HXStr } from '../../src/core/HXStr'
 import { HSymbol } from '../../src/core/HSymbol'
-import { HList } from '../../src/core/HList'
+import { HList } from '../../src/core/list/HList'
 import { HDict } from '../../src/core/dict/HDict'
 import { HVal } from '../../src/core/HVal'
 import { HNa } from '../../src/core/HNa'
