+++ spec/core/ZincReader.spec.ts
@@ -19,7 +19,7 @@ import { HBool } from '../../src/core/HBool'
 import { HXStr } from '../../src/core/HXStr'
 import { HDict } from '../../src/core/dict/HDict'
 import { HSymbol } from '../../src/core/HSymbol'
-import { HList } from '../../src/core/HList'
+import { HList } from '../../src/core/list/HList'
 import { HGrid } from '../../src/core/grid/HGrid'
 import { HVal } from '../../src/core/HVal'
 import { readFile } from '../file'
