+++ spec/core/dict/HDict.spec.ts
@@ -11,7 +11,7 @@ import { Kind } from '../../../src/core/Kind'
 import { HStr } from '../../../src/core/HStr'
 import { HMarker } from '../../../src/core/HMarker'
 import { HBool } from '../../../src/core/HBool'
-import { HList } from '../../../src/core/HList'
+import { HList } from '../../../src/core/list/HList'
 import { HGrid } from '../../../src/core/grid/HGrid'
 import { HVal, TEXT_ENCODER } from '../../../src/core/HVal'
 import { HSymbol } from '../../../src/core/HSymbol'
