+++ spec/core/grid/HGrid.spec.ts
@@ -11,7 +11,7 @@ import { HStr } from '../../../src/core/HStr'
 import { Kind } from '../../../src/core/Kind'
 import { HNum } from '../../../src/core/HNum'
 import { HMarker } from '../../../src/core/HMarker'
-import { HList } from '../../../src/core/HList'
+import { HList } from '../../../src/core/list/HList'
 import { HRef } from '../../../src/core/HRef'
 import { HFilter } from '../../../src/filter/HFilter'
 import { HaysonDict } from '../../../src/core/hayson'
