+++ spec/lodash.spec.ts
@@ -9,7 +9,7 @@ import { forEach, fill } from 'lodash'
 import './matchers'
 import './customMatchers'
 import { HDict, HValRow } from '../src/core/dict/HDict'
-import { HList } from '../src/core/HList'
+import { HList } from '../src/core/list/HList'
 import { HVal } from '../src/core/HVal'
 import { HStr } from '../src/core/HStr'
 import { HNum } from '../src/core/HNum'
