+++ src/core/Array.ts
@@ -11,7 +11,7 @@
  */
 
 import { HGrid } from './grid/HGrid'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HDict } from './dict/HDict'
 import { HVal } from './HVal'
 import { makeValue } from './util'
