+++ src/core/HCoord.ts
@@ -15,7 +15,7 @@ import {
 import { HaysonCoord } from './hayson'
 import { Node } from '../filter/Node'
 import { HGrid } from './grid/HGrid'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HDict } from './dict/HDict'
 import { EvalContext } from '../filter/EvalContext'
 import { JsonV3Coord } from './jsonv3'
