+++ src/core/HRef.ts
@@ -15,7 +15,7 @@ import { HaysonRef } from './hayson'
 import { HStr } from './HStr'
 import { Node } from '../filter/Node'
 import { HGrid } from './grid/HGrid'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HDict } from './dict/HDict'
 import { EvalContext } from '../filter/EvalContext'
 import { JsonV3Ref } from './jsonv3'
