+++ src/core/HRemove.ts
@@ -16,7 +16,7 @@ import {
 import { HaysonRemove } from './hayson'
 import { Node } from '../filter/Node'
 import { HGrid } from './grid/HGrid'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HDict } from './dict/HDict'
 import { EvalContext } from '../filter/EvalContext'
 import { JsonV3Remove } from './jsonv3'
