+++ src/core/HUri.ts
@@ -14,7 +14,7 @@ import {
 import { HaysonUri } from './hayson'
 import { Node } from '../filter/Node'
 import { HGrid } from './grid/HGrid'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HDict } from './dict/HDict'
 import { EvalContext } from '../filter/EvalContext'
 import { memoize } from '../util/memoize'
