+++ src/core/HVal.ts
@@ -6,7 +6,7 @@ import { Kind } from './Kind'
 import { HaysonVal } from './hayson'
 import { HFilter } from '../filter/HFilter'
 import { Node } from '../filter/Node'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HGrid } from './grid/HGrid'
 import { HDict } from './dict/HDict'
 import { EvalContext } from '../filter/EvalContext'
