+++ src/core/HXStr.ts
@@ -15,7 +15,7 @@ import {
 import { HaysonXStr } from './hayson'
 import { Node } from '../filter/Node'
 import { HGrid } from './grid/HGrid'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HDict } from './dict/HDict'
 import { HStr } from './HStr'
 import { EvalContext } from '../filter/EvalContext'
