+++ src/core/ZincReader.ts
@@ -22,7 +22,7 @@ import { HXStr } from './HXStr'
 import { HDict } from './dict/HDict'
 import { HValObj } from './dict/HValObj'
 import { HSymbol } from './HSymbol'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HGrid } from './grid/HGrid'
 import { GridObjStore } from './grid/GridObjStore'
 import { GridColumn } from './grid/GridColumn'
