+++ src/core/dict/HDict.ts
@@ -10,7 +10,7 @@ import { Node } from '../../filter/Node'
 import { HaysonDict, HaysonVal } from '../hayson'
 import { HDateTime } from '../HDateTime'
 import { HGrid } from '../grid/HGrid'
-import { HList } from '../HList'
+import { HList } from '../list/HList'
 import { HMarker } from '../HMarker'
 import { HNamespace, Reflection } from '../HNamespace'
 import { HRemove } from '../HRemove'
