+++ src/core/grid/HGrid.ts
@@ -20,7 +20,7 @@ import { HaysonGrid, HaysonDict } from '../hayson'
 import { HStr } from '../HStr'
 import { HFilter } from '../../filter/HFilter'
 import { Node, isNode } from '../../filter/Node'
-import { HList } from '../HList'
+import { HList } from '../list/HList'
 import { makeValue } from '../util'
 import { HRef } from '../HRef'
 import { EvalContext, EvalContextResolve } from '../../filter/EvalContext'
