+++ src/core/util.ts
@@ -40,7 +40,7 @@ import { HNa } from './HNa'
 import { HCoord } from './HCoord'
 import { HXStr } from './HXStr'
 import { HSymbol } from './HSymbol'
-import { HList } from './HList'
+import { HList } from './list/HList'
 import { HGrid } from './grid/HGrid'
 import { Scanner } from '../util/Scanner'
 import { HNamespace } from './HNamespace'
