+++ src/shorthand.ts
@@ -24,7 +24,7 @@ import { HDateTime } from './core/HDateTime'
 import { HDict } from './core/dict/HDict'
 import { DictStore } from './core/dict/DictStore'
 import { HGrid } from './core/grid/HGrid'
-import { HList } from './core/HList'
+import { HList } from './core/list/HList'
 import { HMarker } from './core/HMarker'
 import { HNa } from './core/HNa'
 import { HNamespace } from './core/HNamespace'
